server/requirements.txt
Django==3.0.1
django-debug-toolbar==2.1
django-filter==2.2.0
djangorestframework==3.11.0
elasticsearch-dsl==7.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.4